New Models, general fixes, code for further generalisation
Experimenting with a form of adaptive delta modulation for better abstraction that only spikes if it goes outside a proportion of the batch's sample deviation and mean. Good generalisation but poor accuracy.

Started changing how I did TAB. Fixes to gesture labels.
Added a potential DANN using a S-LSTM as a classifier- going to add conv compoments for spatial features.