Fitted Abstracted model tothe task
Created a STD and Mean Continual Learning approach for the Ninapro DB6 Dataset with plotting etc. Serves as a base for any statistical loss based TTA with dark experience replay.

Note: I used an old baseline creator to create this and as of such, the code might need to be updated to suit with the newer ones that work on the HPC

Additionally, this code uses the full testLoader with is inefficient. Future work will most likely change this to use one of each type of gesture to adapt to